requirements.txt
colorama==0.4.6
faiss_cpu==1.8.0
Faker==37.1.0
fastapi==0.115.12
Flask==3.1.0
Jinja2==3.1.6
numpy==2.2.4
pandas==2.2.3
Pillow==9.5.0
Pillow==11.2.0
plotly==5.22.0
Requests==2.32.3
rich==14.0.0
scikit_learn==1.4.2
sentence_transformers==3.2.1
uvicorn==0.34.0
waitress==3.0.1
